moved sequelize variable to config/sequelize.js
