+++ src/db/models/index.js
@@ -2,22 +2,11 @@
 
 var fs = require('fs');
 var path = require('path');
-var Sequelize = require('sequelize');
 var basename = path.basename(__filename);
-var env = process.env.NODE_ENV || 'development';
-var config = require(__dirname + '/../config/config.js')[env];
-var db = {};
+var Sequelize = require('sequelize');
 
-if (config.use_env_variable) {
-  var sequelize = new Sequelize(process.env[config.use_env_variable], config);
-} else {
-  var sequelize = new Sequelize(
-    config.database,
-    config.username,
-    config.password,
-    config
-  );
-}
+var sequelize = require('../config/sequelize');
+var db = {};
 
 fs.readdirSync(__dirname)
   .filter(file => {
